+++ kind-rational/CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 0.3.1
+
+* COMPILER ASSISTED BREAKING CHANGE: `rationalVal`, `someRationalVal` and
+  `fromSRational` now deal with `KindRational`'s `Rational`s, rather than
+  `Prelude`'s `Rational`s.
+
+* COMPILER ASSISTED BREAKING CHANGE: Removed `fromSRational'`.
+
+* Added `SingI` and `SingKind` instances.
+
+
 # Version 0.3
 
 * COMPILER ASSISTED BREAKING CHANGE: `TestEquality` and `TestCoercion`